Move common logic into own SongProgressBar class
