+++ osu.Game/Screens/Play/HUD/DefaultSongProgressBar.cs
@@ -7,67 +7,23 @@ using osuTK.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Utils;
-using osu.Framework.Threading;
 
 namespace osu.Game.Screens.Play.HUD
 {
-    public partial class DefaultSongProgressBar : SliderBar<double>
+    public partial class DefaultSongProgressBar : SongProgressBar
     {
-        /// <summary>
-        /// Action which is invoked when a seek is requested, with the proposed millisecond value for the seek operation.
-        /// </summary>
-        public Action<double>? OnSeek { get; set; }
-
-        /// <summary>
-        /// Whether the progress bar should allow interaction, ie. to perform seek operations.
-        /// </summary>
-        public bool Interactive
-        {
-            get => showHandle;
-            set
-            {
-                if (value == showHandle)
-                    return;
-
-                showHandle = value;
-
-                handleBase.FadeTo(showHandle ? 1 : 0, 200);
-            }
-        }
-
         public Color4 FillColour
         {
             set => fill.Colour = value;
-        }
-
-        public double StartTime
-        {
-            set => CurrentNumber.MinValue = value;
-        }
-
-        public double EndTime
-        {
-            set => CurrentNumber.MaxValue = value;
-        }
-
-        public double CurrentTime
-        {
-            set => CurrentNumber.Value = value;
         }
 
         private readonly Box fill;
         private readonly Container handleBase;
         private readonly Container handleContainer;
 
-        private bool showHandle;
-
         public DefaultSongProgressBar(float barHeight, float handleBarHeight, Vector2 handleSize)
         {
-            CurrentNumber.MinValue = 0;
-            CurrentNumber.MaxValue = 1;
-
             RelativeSizeAxes = Axes.X;
             Height = barHeight + handleBarHeight + handleSize.Y;
 
@@ -130,9 +86,11 @@ namespace osu.Game.Screens.Play.HUD
             };
         }
 
-        protected override void UpdateValue(float value)
+        protected override void LoadComplete()
         {
-            // handled in update
+            base.LoadComplete();
+
+            InteractiveBindable.BindValueChanged(i => handleBase.FadeTo(i.NewValue ? 1 : 0, 200), true);
         }
 
         protected override void Update()
@@ -144,18 +102,6 @@ namespace osu.Game.Screens.Play.HUD
 
             fill.Width = newX;
             handleBase.X = newX;
-        }
-
-        private ScheduledDelegate? scheduledSeek;
-
-        protected override void OnUserChange(double value)
-        {
-            scheduledSeek?.Cancel();
-            scheduledSeek = Schedule(() =>
-            {
-                if (showHandle)
-                    OnSeek?.Invoke(value);
-            });
         }
     }
 }
